I cannot lend PragProg ebooks :(
http://pragprog.com/frequently-asked-questions/ebooks

Are the eBook files restricted?

There is no copy protection or functionality restrictions in our eBooks. You may view or print them for personal use as you see fit.

You may not give your PDF, epub or mobi versions to other people. The file you order is personalized with your name and other identifying information.

You can buy multiple licenses of an eBook title for your team or organization, in which case the eBook will be stamped with the number of allowed licenses. We’ll only send you one, so as to conserve everyone’s bandwidth.